update cursor positions when indenting
This change causes cursor(s) to maintain their relative position
in their lines after the (in/out)dent occurs.

The previous behaviour was for the whole line to be selected
if nothing was selected at the start.  This was frustrating
because it does not let you immediately start typing after indenting
a new line.